Got email alerts to work with traffic and weather
I moved the email alert functionality into a new file that exports the
email function out to other files through the function sendEmail.

This gets called by the traffic and weather functions when they know
they have a new alert. Since both weather and traffic have differrent
functinos the new alert detection method was implemented differently.
